requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
Django==5.2.6
django-cors-headers==4.9.0
django-ninja==1.4.3
django-tailwind==4.2.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.15.0
opencv-python 
moviepy
boto3==1.35.0
django-storages==1.14.4
python-dotenv==1.0.0
psycopg2-binary==2.9.9
reportlab==4.4.5
weasyprint==66.0
celery==5.4.0
redis==5.0.1
django-celery-results==2.5.1
firebase-admin==6.5.0
